+++ CHANGELOG.md
@@ -1,5 +1,6 @@
 Version 8.0.0-dev
 - dropped support for Composer 2.0 and 2.1
+- BC break: removed support for old way of providing 1 parameter to a test method via data provided
 
 Version 7.3.1
 - allowed installation konecnyjakub/event-dispatcher 2